requirements/requirements_dev.txt
# This file was autogenerated by uv via the following command:
#    uv export --frozen --no-hashes --python 3.11
-e .
-e ./packages/_dev
-e ./packages/boilercine
-e ./packages/boilercore
-e ./packages/context_models
-e ./packages/pipeline
./packages/cachier
accessible-pygments==0.0.5
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
alabaster==0.7.16
amqp==5.2.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.5.0
appdirs==1.4.4
appnope==0.1.4 ; platform_system == 'Darwin'
argcomplete==3.5.0
asciitree==0.3.3
astroid==3.3.1
asttokens==2.4.1
asyncssh==2.17.0
atpublic==5.0
attrs==24.2.0
av==13.0.0
babel==2.16.0
backoff==2.2.1
beautifulsoup4==4.12.3
billiard==4.2.0
bleach==6.1.0
blosc2==2.7.1
bokeh==3.5.2
bottleneck==1.4.0
cachetools==5.5.0
cappa==0.22.5
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
cftime==1.6.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cligj==0.7.2
cloudpickle==3.0.0
colorama==0.4.6
colorcet==3.0.1
comm==0.2.2
configobj==5.0.8
contourpy==1.3.0
copykitten==1.1.1
coverage==7.6.1
cryptography==43.0.1
cycler==0.12.1
cyclopts==2.9.9
dask==2024.9.0
dask-expr==1.1.14
debuglater==1.4.4
debugpy==1.8.2
decopatch==1.4.10
decorator==5.1.1
defusedxml==0.7.1
devtools==0.12.2
dictdiffer==0.9.0
dill==0.3.8
diskcache==5.6.3
distlib==0.3.8
distributed==2024.9.0
distro==1.9.0
docopt==0.6.2
docstring-parser==0.16 ; python_full_version < '4.0.0'
docutils==0.21.2
dpath==2.2.0
dulwich==0.22.1
dvc==3.55.2
dvc-data==3.16.6
dvc-gs==3.0.1
dvc-http==2.32.0
dvc-objects==5.1.0
dvc-render==1.0.2
dvc-studio-client==0.21.0
dvc-task==0.4.0
entrypoints==0.4
execnet==2.1.1
executing==2.1.0
fasteners==0.19 ; sys_platform != 'emscripten'
fastjsonschema==2.20.0
fawltydeps==0.16.0
filelock==3.16.1
fiona==1.10.1
flatten-dict==0.4.2
flexcache==0.3
flexparser==0.3.1
flox==0.9.12
flufl-lock==8.1.0
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.9.0
funcy==2.0
future==1.0.0
gcsfs==2024.9.0.post1
geopandas==0.14.3
gitdb==4.0.11
gitpython==3.1.43
gmpy2==2.2.0rc1
google-api-core==2.20.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grandalf==0.8
greenlet==3.1.0 ; (python_full_version < '3.13' and platform_machine == 'AMD64') or (python_full_version < '3.13' and platform_machine == 'WIN32') or (python_full_version < '3.13' and platform_machine == 'aarch64') or (python_full_version < '3.13' and platform_machine == 'amd64') or (python_full_version < '3.13' and platform_machine == 'ppc64le') or (python_full_version < '3.13' and platform_machine == 'win32') or (python_full_version < '3.13' and platform_machine == 'x86_64')
gto==1.7.1
h11==0.14.0
h5netcdf==1.3.0
h5py==3.11.0
hydra-core==1.3.2
identify==2.6.1
idna==3.10
imageio==2.34.1
imagesize==1.4.1
importlib-metadata==8.5.0
incremental==24.7.2
iniconfig==2.0.0
ipykernel==6.29.4
ipython==8.26.0
isort==5.13.2
iterative-telemetry==0.0.9
jedi==0.19.1
jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-cache==1.0.0
jupyter-client==8.6.3
jupyter-core==5.7.2
jupyterlab-pygments==0.3.0
kiwisolver==1.4.7
kombu==5.4.2
latexcodec==3.0.0
lazy-loader==0.4
linkify-it-py==2.0.3
llvmlite==0.43.0
locket==1.0.0
loguru==0.7.0
looseversion==1.3.0
lz4==4.3.3
makefun==1.15.4
markdown-it-py==3.0.0
markupsafe==2.1.5
matplotlib==3.7.2
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
mistune==3.0.2
monotonic==1.6
more-itertools==10.4.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
myst-nb==1.1.1
myst-parser==3.0.1
nbclient==0.10.0
nbconvert==7.7.3
nbformat==5.10.4
ndindex==1.8
nest-asyncio==1.6.0
netcdf4==1.6.5
networkx==3.3
nodeenv==1.9.1
nptyping==2.5.0
numba==0.60.0
numbagg==0.8.1
numcodecs==0.13.0
numexpr==2.10.1
numpy==1.26.0
numpy-groupies==0.11.2
numpydantic==1.0.0
numpydoc==1.7.0
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.10.0.84
orjson==3.10.7 ; implementation_name == 'cpython'
packaging==24.1
pandas==2.2.2
pandas-stubs==2.2.2.240514
pandocfilters==1.5.1
param==2.1.1
parso==0.8.4
partd==1.4.2
pathspec==0.12.1
pexpect==4.9.0 ; sys_platform != 'emscripten' and sys_platform != 'win32'
pillow==10.3.0
pint==0.24.3
pip==24.2
pip-requirements-parser==32.0.1
pipdeptree==2.23.0
pipx==1.6.0
platformdirs==3.11.0
ploomber-core==0.2.25
ploomber-engine==0.0.30
pluggy==1.5.0
pooch==1.8.2
portalocker==2.3.2
posthog==3.6.6
pre-commit==3.7.1
prompt-toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.28.2
psutil==6.0.0
ptyprocess==0.7.0 ; sys_platform != 'emscripten' and sys_platform != 'win32'
pure-eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==14.0.1
pyasn1==0.6.1
pyasn1-modules==0.4.1
pybtex==0.24.0
pybtex-docutils==1.0.3
pycine @ git+https://github.com/ottomatic-io/pycine@815cfca06cafc50745a43b2cd0168982225c6dca
pycparser==2.22
pyct==0.5.0
pydantic==2.9.1
pydantic-core==2.23.3
pydantic-settings==2.5.2
pydata-sphinx-theme==0.15.4
pydot==3.0.1
pygit2==1.15.1
pygments==2.18.0
pygtrie==2.5.0
pyparsing==3.0.9
pyproj==3.6.1
pyqtgraph==0.13.3
pyright==1.1.371
pyside6==6.7.2
pyside6-addons==6.7.2
pyside6-essentials==6.7.2
pytest==8.2.2
pytest-cov==5.0.0
pytest-custom-exit-code==0.3.0
pytest-github-actions-annotate-failures==0.2.0
pytest-harvest==1.10.5
pytest-plt==1.1.1
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2023.3
pywin32==306 ; sys_platform == 'win32' or platform_system == 'Windows'
pyyaml==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.8.1
rich-rst==1.3.1
rpds-py==0.20.0
rsa==4.9
ruamel-yaml==0.18.6
ruamel-yaml-clib==0.2.8 ; python_full_version < '3.13' and platform_python_implementation == 'CPython'
ruff==0.5.1
scikit-image==0.23.1
scipy==1.11.2
scmrepo==3.3.7
seaborn==0.13.2
semver==3.0.2
setproctitle==1.3.3
setuptools==70.0.0
shapely==2.0.3
shellingham==1.5.4
shiboken6==6.7.2
shortuuid==1.0.13
shtab==1.7.1
six==1.16.0
smmap==5.0.1
snakeviz==2.2.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.6
sourcery==1.21.0
sparklines==0.4.2
sphinx==7.3.7
sphinx-autobuild==2024.4.16
sphinx-autodoc2==0.5.0
sphinx-book-theme==1.1.3
sphinx-design==0.6.0
sphinx-thebe==0.3.1
sphinx-tippy==0.4.3
sphinx-togglebutton==0.3.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-bibtex==2.6.2
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-mermaid==0.9.2
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sphinxcontrib-towncrier==0.4.0a0
sqlalchemy==2.0.35
sqltrie==0.11.1
stack-data==0.6.3
starlette==0.38.5
stopit==1.1.2
sympy==1.12
tables==3.10.1
tabulate==0.9.0
tblib==3.0.0
tifffile==2024.8.30
timecode==1.4.0
tinycss2==1.3.0
tomli==2.0.1 ; python_full_version <= '3.11'
tomlkit==0.12.4
toolz==0.12.1
tornado==6.4.1
towncrier==23.11.0
tqdm==4.66.3
trackpy==0.6.2
traitlets==5.14.3
typer==0.12.5
types-pytz==2024.2.0.20240913
typing-extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.1
uc-micro-py==1.0.3
uncertainties==3.1.7
urllib3==2.2.2
userpath==1.9.2
uvicorn==0.30.6
vine==5.1.0
virtualenv==20.26.5
voluptuous==0.15.2
watchdog==2.3.1
watchfiles==0.22.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==13.0.1
wheel==0.44.0
win32-setctime==1.1.0 ; sys_platform == 'win32'
xarray==2023.7.0
xyzservices==2024.9.0
yarl==1.11.1
zarr==2.18.3
zc-lockfile==3.0.post1
zict==3.0.0
zipp==3.20.2
